[BE] refactor : dto 패키지 변경 및 ServiceDtoConverter 추가
