+++ backend/reviewduck/src/main/java/com/reviewduck/template/controller/TemplateController.java
@@ -25,12 +25,12 @@ import com.reviewduck.review.dto.request.ReviewFormCreateRequest;
 import com.reviewduck.review.dto.response.ReviewFormCodeResponse;
 import com.reviewduck.review.service.ReviewFormService;
 import com.reviewduck.template.domain.Template;
-import com.reviewduck.template.dto.request.TemplateCreateRequest;
-import com.reviewduck.template.dto.request.TemplateUpdateRequest;
-import com.reviewduck.template.dto.response.MemberTemplatesResponse;
-import com.reviewduck.template.dto.response.TemplateIdResponse;
-import com.reviewduck.template.dto.response.TemplateResponse;
-import com.reviewduck.template.dto.response.TemplatesResponse;
+import com.reviewduck.template.dto.controller.request.TemplateCreateRequest;
+import com.reviewduck.template.dto.controller.request.TemplateUpdateRequest;
+import com.reviewduck.template.dto.controller.response.MemberTemplatesResponse;
+import com.reviewduck.template.dto.controller.response.TemplateIdResponse;
+import com.reviewduck.template.dto.controller.response.TemplateResponse;
+import com.reviewduck.template.dto.controller.response.TemplatesResponse;
 import com.reviewduck.template.service.TemplateService;
 
 import io.swagger.v3.oas.annotations.Operation;
