+++ backend/reviewduck/src/main/java/com/reviewduck/template/dto/controller/request/TemplateUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.reviewduck.template.dto.request;
+package com.reviewduck.template.dto.controller.request;
 
 import java.util.List;
