+++ backend/reviewduck/src/main/java/com/reviewduck/template/dto/controller/response/MemberTemplateResponse.java
@@ -1,9 +1,8 @@
-package com.reviewduck.template.dto.response;
+package com.reviewduck.template.dto.controller.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.reviewduck.member.domain.Member;
 import com.reviewduck.template.domain.Template;
 
 import lombok.AccessLevel;
@@ -12,27 +11,19 @@ import lombok.Getter;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class TemplateResponse {
+public class MemberTemplateResponse {
 
-    private boolean isCreator;
     private TemplateInfoResponse info;
-    private CreatorResponse creator;
     private List<TemplateQuestionResponse> questions;
 
-    public static TemplateResponse of(Template template, Member member) {
+    public static MemberTemplateResponse from(Template template) {
         List<TemplateQuestionResponse> questions = template.getQuestions().stream()
             .map(TemplateQuestionResponse::from)
             .collect(Collectors.toUnmodifiableList());
 
-        return new TemplateResponse(
-            template.isMine(member),
+        return new MemberTemplateResponse(
             TemplateInfoResponse.from(template),
-            CreatorResponse.from(template.getMember()),
             questions
         );
-    }
-
-    public boolean getIsCreator() {
-        return isCreator;
     }
 }
