+++ backend/reviewduck/src/main/java/com/reviewduck/template/dto/controller/response/TemplateResponse.java
@@ -1,8 +1,9 @@
-package com.reviewduck.template.dto.response;
+package com.reviewduck.template.dto.controller.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.reviewduck.member.domain.Member;
 import com.reviewduck.template.domain.Template;
 
 import lombok.AccessLevel;
@@ -11,19 +12,27 @@ import lombok.Getter;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
-public class MemberTemplateResponse {
+public class TemplateResponse {
 
+    private boolean isCreator;
     private TemplateInfoResponse info;
+    private CreatorResponse creator;
     private List<TemplateQuestionResponse> questions;
 
-    public static MemberTemplateResponse from(Template template) {
+    public static TemplateResponse of(Template template, Member member) {
         List<TemplateQuestionResponse> questions = template.getQuestions().stream()
             .map(TemplateQuestionResponse::from)
             .collect(Collectors.toUnmodifiableList());
 
-        return new MemberTemplateResponse(
+        return new TemplateResponse(
+            template.isMine(member),
             TemplateInfoResponse.from(template),
+            CreatorResponse.from(template.getMember()),
             questions
         );
+    }
+
+    public boolean getIsCreator() {
+        return isCreator;
     }
 }
