+++ backend/reviewduck/src/main/java/com/reviewduck/template/dto/controller/response/TemplateSummaryResponse.java
@@ -1,4 +1,4 @@
-package com.reviewduck.template.dto.response;
+package com.reviewduck.template.dto.controller.response;
 
 import com.reviewduck.member.domain.Member;
 import com.reviewduck.template.domain.Template;
